Fix quotes in Opera. #388.
We're not sending an encoding header for JavaScript, and, even if we were,
Opera is reputed not to like it. Most browsers are interpreting the file as
UTF-8, but Opera is interpreting it as ISO-8859-1 (which is correct, if
annoying). Instead, let's encode the few characters we want to send as escape
sequences.

It might be better to put this into the template (<script type="text/plain">)
and read it from there, but this should fix things for now.