Remove artefacts.js and chosen jquery plugin.
The artefacts.js appears to be redundant.  It's not required by any of
the templates and it now appears to live in panopticon (commit
ec72642385230804b523f6a9e6124184e03cc9ce suggests that it was introduced
for panopticon in the first place).

The artefacts.js file was the only thing using the chosen jquery plugin
so I've removed that oo.