Add Google Analytics tracking for Guide success.
As the page is loading an custom event is fired to say that it was entered.
On DOM ready the success tracking function is set up. This sends custom
events for Success, Success due to being on the site for more than seven
seconds and Success due to clicking to another page on the guide.