Use min-height for header bar in modern browsers
Switching from height to min-height doesn't have any effect on full page
width documents. However, when a document is responsive (as inside
government are) it will enable the black bar to grow to ensure it always
covers the contents of the bar. Otherwise you can get the white links on
a white background.

I have forced old IE to still use height as in testing it will otherwise
change the height of the heading bar.