Update text.scss to take advantage of SASS
As part of the frontend toolkit work that @intranation has done, we
should be merging this in and getting as much of our code using it as we
can.

This change simplifies a lot of the code that we have for article text
by structuring it to use sass blocks, helpers, and our own mixin
methods.