Update publisher stylesheets to use SCSS
As part of the move to SCSS, update all of these styles accordingly
using helpers/mixings from the frontend toolkit.

There is one caveat in this file: whitespace. At line 442 there is an
extra space put in on purpose to make sure that the block is readible
for anyone who may come across it. As always this is up for debate, but
was an initial decision made by @tombye, @edds, and myself.