Fix the overflow bug in IE7
Due to the article height setting of 30em to deal with the min-height
hack it keeps the bounding box at a fixed height.

In cases where the article tag has the .group class applied it's because
the inner content extends past this enforced height restriction. Remove
the height restriction in these cases to fix the IE7 rendering bug.