Update feedback options
A new feedback form is on the way and will replace
the existing methods of feeding back to GDS about
GOV.UK

We've also introduced a new 'report a problem'
form onto every piece of content that means links
to the existing page are not as useful anymore.

Links will be reintroduced once the new form is
available.