Added eventFreeze method to jQuery, added tests
Added a method to the jQuery namespace for freezing and resuming
events, creates a minimal copy of the event on ie6,7,8 and fully
copies the event on ie9, chrome, firefox. Anyone consuming events
outside of analytics.js shouldn't notice that the event was ever
touched.

For tests see event-ice-bucketSpec.js
   -- @gtrogers