Fixed entry/success tracking for smart answers
Generally we're not supposed to track entry if a user navigates within
a format. In case of smart answers success can only happen after
user navigated to a different page than initial. Because of that
we need to kind of "bind" these two visits together to be logically
considered as one.