Exit early from tab builder if tabs aren't in HTML
This was causing the search page to die because we don't show any tabs
on it at all if only one index gets hit.

Long term it should be smarter about not hiding content if it's the only
thing to show, but this was the quickest and easiest way to fix the bug.