Make sure we don't minify the copyright message
As part of the release of gov.uk we need the copyright message to stay
in all files that we use to ship the fonts to browsers.

Adding an inline comment (which will be removed) to make it clear why
the specific formatting of the comment has been used for the copyright
message.