Make related sticky scroll work with tabbed content
Tabbed content pages will be taller at load than when tabs are applied
as only the selected content is shown. The tabs JS therefore now calls a
method on the scrolling object to update its value for the page height.

This method was added with version 0.7.1 of the toolkit in this commit:

https://github.com/alphagov/govuk_frontend_toolkit/commit/6d9e1