Add border-top hack for input buttons
Inputs cannot hold child elements so pseudo-classes will not help
increase the click target. Inputs do not seem to have the problem with
the space below their text being unclickable.

The remaining problem is that the button moves away from the cursor when
clicked at the top. This adds a border-top to catch this.