Don't need a cookie for seeing the bar
Showing the bar will occur randomly. We don't need to track this at
the moment in the form of cookies. Just remove it as the only thing we
need is the cookie to check whether a user has 'taken the survey'.

In this case, 'taken the survey' means that the user has either taken
the survey or chosen to hit 'no thanks'. In either case, setting the
cookie of the same name means that the user satisfaction bar will no
longer exist.