Add pixel ratio density tracking in Google Analytics
I've added this to the _google_analytics partial that gets included in the main template because I couldn't find a relevant place to add it in the analytics javascript files are they tracking event success.

I did try adding this to core.js but that would have required another page tracking call to GA which I was worried may have impacted our analytics. Custom Variables must be set before any tracking calls, and as this is where the main tracking call is made, it made sense to me to add it here.

If this is incorrect or there is a better home for this please let me know.