Add GA tracking __utm.gif to print styles
This JS constructs a __utm.gif URL and adds it to a CSS print media query which is inserted into the `<head>` of each page. When the print stylesheets are loaded (on printing a page) the gif URL is hit causing GA to record a visit.

We need to check that the visits are being recorded. 

* Are these the only required arguments for GA?
* Are the values in the arguments enough for GA to record the metrics correctly?