Add chromeless wrapper template
We need a chromeless wrapper template for html-publications. However,
these should still retain the actual footer, the cookie notice and the
analytics tracking. I have tried to split out the core parts of the
template that this shares with base so there is little repitition.

I also noticed a bug where if you set a custom css filename you wouldn't
get IE fallback stylesheets and would just get the moden browser version
for all IE's. This fixes that issue also.