Redirect requests for /apple-touch* to the assets
Seems like iOS (and Android) request /apple-touch-icon.png and its
variants, resulting in a pile of 404s on the main site.

This is explicitly against Apple's own documentation, but as the logs
don't lie we should serve a 301 to the right place here. The 301s will
get cached and stop bothering us.