Simplify jQuery usage to automatically filter elements it can't find
This block of code should operate on either of the report-a-problem
forms that exist. They won't be present at the same time, but we were
using a wrapped form to append across them.

If an element "#foo" exists on the page but the element "#bar"
doesn't, running $("#foo, #bar") will only return an array containing
a single jQuery object with "#foo".

This means we can reduce our error handling for a given page.