Do not show the satisfaction survey on transaction done pages
If users complete the survey from a done page there is the worry that
they will fill out their response with the context of a given
departmental transaction rather than the GOV.UK design.

Additionally, multi-variate testing it currently underway on one of
the transaction pages so we don't want to mess with the results
gathering while this is happening.