Configure asset pipeline not to compile assets.
This will cause it to fail noisily if we reference a non-existent asset,
or one that's not been added to the precompile list.  Setting this would
have meant that the problem fixed in #288 would have been noticed
immediately.

These are the Rails default options.  Looking at the git history, it
would seem that they were changed early on in the development while
stuff was being figured out, and then never set back.