Exit tracking uses slug in data attribute if available.
This change makes exit tracking use the slug in the
data-transaction-slug attribute on the link if it's available.  If not
it will fall back to parsing the URL.

This will fix a number of issues we're seeing with exit tracking where
it's not able to get the correct slug from the URL (e.g. mixed case
slugs, pages from the google cache etc...)