Add YouTube account verification file
We need to validate our YouTube account so we can embed deep links in
the annotations.

The primary use case for this is linking out to help or transaction
start pages from instructional videos hosted on YouTube, e.g link to the
passport application process from a video about the transaction.