Update footer CSS to be the same width as header
The footer was wider than the header on anything below 1024 width. This
changes the markup so it now matches the header grid and resizes at the
same points as the header and the copyright notice below it.

Moved to two floated elements of 1/3rd and 2/3rd widths respectivly
rather than elements close to those widths.

Updated code to use the `media` mixin rather than the `media-down` which
meant we were able to reduce a large amount of code.

Added a `<hr>` for the break between the links and the copyright as
couldn't use a border on any elements as they were `15px` to wide on
either side due to them containing floated elements which managed their
own gutter.