Move banner_notification into after_header block
This moves the banner_notification from the top of the content
block and into the bottom of the after_header block. This
has the effect of moving the campaign bar above the global
header bar - it looked odd when it was below it.

Arguably the banner notification should be above the user satisfaction
survey because it's more important, but I didn't want to make that
change now.