Track when we show the browser upgrade prompt
At the moment it's getting quite a lot of negative feedback, and knowing who actually
sees it will help use decide how to respond to that.

We've been inferring this from traffic to the /help/browsers page, but that only captures
users who are clicking on the banner. This will record exactlty who sees it.