User satisfaction markup injected by survey script
Previously, the markup was included in the base template all the time.
Apart from being an extra bit of markup in every request, this was also
causing issues with unusual clients, search spiders, etc.

This leaves only a wrapper with no content in the base template, and puts
the markup into the survey script. Since this script is going to be cached
by clients.