Serve govuk-component documentation as JSON
Previously this documentation was stored in the govuk_component_guide app and was
frequently out of sync with static, or failed if pointing at a different static
endpoint.

This moves the documentation into static and exposes it as an API-like endpoint,
so that the documentation data could be auto generated in the future, rather than
a static file.