Add a rails generator for GOV.UK components
Generates all the file scaffolding, in the right places, as
the files are a bit spread throughout static at the moment.

This will include all the files required to let the component
be demo'd on govuk-component

Appending to the json docs file is a bit shonky, but we're
not planning on using that as a docs method long term. It'll
likely be inline with template, or at least one doc file per
component, which will make this simpler.