Move selectable Sass into header footer stylesheet
Ideally this would be a component with matching markup. However, the
JavaScript to power this and applications currently using this would
need updating to match a component namespace.

As the JavaScript is already served in the header_footer_only JavaScript
file, move the Sass into header_footer_only as well.

Also better scope the Sass during the move so that it is all nested
within the `label.selectable` selector. This will enable switching the
code to a proper component much easier in the future.