Add Google Analytics tracking for events on the Satisfaction Survey
To better understand how our users interact with the user satisfaction
survey, push some tracking events to Google Analytics. With this
change we'll be able to count the different interactions and get a
better understanding (in the specific) of what's occurring beyond the
case where a user has taken the survey.

It's worth noting that the tracked events are set as non-interactive
events in Google Analytics. A description of this is here:
https://developers.google.com/analytics/devguides/collection/gajs/eventTrackerGuide#non-interaction