Put the 'mainstream' footer blue bar on #footer not #wrapper
This is so we can use the new govuk_frontend_toolkit grid
helpers directly on #wrapper, without the 'footer' blue
bar also being constrainted to the grid container width.

By moving the blue bar to #footer, it will always be 100%
width, maintaining the current design.

There is one visual regression, in the current design
the 10px blue bar is applied above the footers 1px grey
bar, but in the new version the blue bar replaces it
entirely.

I debated doing some selector gymnastics to keep the 1px
bar as well, but after a discuss with Mark H. we agreed
that is was barely percievable and fine to be replaced.