__canary__ is now in Puppet
Since https://github.gds/gds/puppet/pull/2288, the __canary__ has been in
the Nginx configuration in Puppet. We don't need this file as well as that
one, as the Nginx configuration is serving the content (and setting it to be
JSON, much like the __canary__ on GOV.UK).

Removes the file to avoid confusion. Cc: @dancarley