Add govspeak component
Create a component for govspeak. All this will do is output the content
passed to it. By being a component it means that the styling can be
managed in a consistent place, so changes can be rolled out to all of
our text blocks in one go.

The Sass for this has mostly been lifted straight out of whitehall with
some minor alterations (things like using standard typography mixins
rather than whitehall specific ones).