Add title component
Add a component for page titles. The title should be large by default
with an option to make it smaller for longer titles.

On average the titles on government bits of content are twice as long as
those found on general guidance. We shrink the font size down so that
90% of titles will fit onto two lines.