Enhance YouTube videos in govspeak content
Govspeak that contains links to youtube videos should be expanded into
an accessible video player. This adds the styles for the video player
and adds the JavaScript to identify the appropriate youtube links and
turn them into video players.

Moved the includes for jQueryUI and the player jQuery plugin into the
govspeak video component JavaScript file. This ensures that all the
dependencies are loaded.