Update jQuery UI to match the version in whitehall
The version of jQuery UI in this repo and the version in whitehall were
incompatible. This causes all the JavaScript on whitehall pages to die.
As a quick fix update the version of jQuery UI to match the version in
whitehall. This will mean users will be downloading two identical
versions of jQuery UI but it means they won't break each other.

As a longer term fix we will remove jQuery UI and the media player from
whitehall. That should be happening as part of the work to remove leaf
pages from whitehall as they are the only things which use videos.