Move footer list item width into media query
On mobile devices this was causing the page to be wider than the width
of the viewport. This was due to the element having 100% width with
padding on the element. Moving the width into a media query will mean
the 100% is only applied on desktop which is where it is needed as it is
overriding the 50% width the other footer list items need.