Allow columns to be flipped for RTL content
For right to left locales the reading edge is on the other side of the
page. In reality this means that the most important content should be
against the right side of the page and the columns should be flipped.
This makes it easy to have all the columns on the page flipped by
adding a wrapping class to the page.