Remove old analytics JavaScript
This analytics was added for the performance platform to create an
engagement score of pages. They have since stopped using these events
and they are no longer needed. Remove them so they stop sending lots of
events to Google Analytics.

Tidy up the function in `tracking.js` which is still needed so that it
is clearer as to what is going on.