Make content and related float next to each other
Historically the related links in the side bar would scroll with the
page. This was removed so the related links are now static on the page.

Update the styles to use floated columns for the sidebar and the main
content. Use the grid mixin from the govuk_toolkit to let this happen.

Remove a collection of CSS that doesn't make sense when the containing
`main` is defining the layout rather than each element on the page
defining how wide they should be.