Improve keyboard accessibility
Though the option select can gain keyboard focus (because it has `tabindex=0`) tabbing past it results in tabbing through all of the checkboxes, even if they are hidden.

_Keyboard users vs screen reader users_
If I were to fix this just for keyboard users, I would make the checkboxes invisible unless you had "opened" the option-select so you couldn't possible tab through them unless they were on the screen i.e. you had opened the option-select. However the way these option-selects currently work with screen readers is they just present as a list of checkboxes, meaning screen-reader users can interact with them as they would a normal list of checkboxes. I like this solution as it is fairly future proof and doesn't rely on JavaScript at all. If we change the tabbing behaviour for sighted keyboard users we might end-up breaking the interaction for screen-reader users.

The implementation here strikes a middle-ground that I'm still not happy with. It is as follows:
- Option-select gains focus
- User presses `tab` key again and the option-select opens
- User now has to `tab` through all of the checkboxes

It seems likely that there is a better way than this, so if anyone has any suggestions for either a solution or better compromise please comment.

This commit contains:
- JavaScript for listening to keyboard events
- Tests for keyboard event listener js