Pass default value to GA httpStatusCode variable.
We track the HTTP status code using the GA custom variable
`httpStatusCode` for all 4xx and 5xx responses. However, it's
remarkably difficult to filter for "custom variable not set"
in GA [1], and actually we're usually interested in getting
all successful responses, rather than unsuccessful ones.

So, if this is not a 4xx or 5xx response, pass a default value
of 200 to GA, to simplify filtering. I've confirmed with @edds
that it's safe to assume that the response is 200 if
`window.httpStatusCode` is unset at this point.

[1] http://stackoverflow.com/questions/7972978/google-analytics-exclude-empty-custom-variable-in-a-custom-report