Fix Last updated on Done page
Last updated paragraph is floated left causing the first word to display inline with the feedback form. All styles for the paragraph are generic styles for '.meta-data p' which could affect several apps in many places. So I scoped this simple fix to '.meta-data p' that also have the .modified-date class.

https://www.pivotaltracker.com/story/show/87918506