Fix appearance of tabs within content-block
When we changed the default styling of text content to use an element
with a class of `content-block` rather than just using an article we
increased the specificity of the CSS selector. This means that the
standard list styling was overriding some of the styles which were being
added by the tab styles. This adds an extra selector to the tabs styles
so it now has a greater specificity than the standard list styling.

I have added an extra selector rather than just changing the selectors
because due to the way the JavaScript works it does checks on the styles
of the elements before it adds in the extra classes to work out if the
tabs should be shown as tabs or an accordion.