Add a component for /government navigation
This commit adds a component for the government_navigation which
contains only markup. This can be included anywhere on GOV.UK and will
be lifted by Slimmer and placed into the header, appropriately styling
it.

It can be passed a string assigned to active which will then mark the
matching link as active. See the code for which strings are used.

This component is slightly odd as Slimmer will move it from the page the
first instance it is included and it doesn't come with it's own CSS so
the first example on the component view in the guide contains nothing
and the second example contains an unstyled list.