Set height to give visual cue of scrollability
On certain combinations of browsers/operating system/number of lines
per option it isn't clear that option-select is scrollable.

This is because the `max-height` of the container lines up with
the bottom edge of the container.

This commit adjusts the height of the container so that the first
line of the last visible option is always cut exactly in half
horizontally.