Add Analytics dimensions for History Mode
Documents in Whitehall that are political and published
by a previous government will be marked as historical.

This adds two new dimensions that will be used to track
content related to History mode, but they've been kept
a little more general, so they could be applied more
widely in the future.

The new dimensions are:
 - PublishingGovernment, the government a document was
   first published under (which may not be when it was
   first published to GOV.UK)
 - PoliticalStatus, whether a document is historic
   (political and non current), political (political
   and current) or apolitical