Fix failing test around option select
The Jasmine tests were failing on the command line with the following error:
```
GOVUK.OptionSelect setupHeight expands the checkbox-container just enough to cut the last visible item in half horizontally, if there are many items
  Expected 29 to be greater than 100.
    Error: Expected 29 to be greater than 100.
```

This test was, however, passing in the browser test runner.

This inconsistency is because of a difference between viewport sizes in the
browser and command line test runners. This commit addresses this inconsistency
by explicitly setting the height and width of the checkbox container in the
failing test.