On option select, only set height one time
Browsers maintain the scroll position of an option select even when they are
closed. This confuses the code that tries to calculate the height of the
container to cut off the last 'visible' item.

For some scroll positions this causes the container to get progressively
shorter each time it is opened.

To fix this the option select should only have its height set once, the first
time it is opened.