Remove 'clearing link' from option-select
Ahead of some accessibility changes that Léonie has recommended, this
commit removes the 'clear' link at the top of the option-select.

The future work is to change the whole clickable label div into a button
which properly exposes/hides the checkbox list for screen readers and
will greatly improve the keyboard accessibility of the widget too.

Making this top bit a button means there is no obvious place for the
clear link to go, meaning it has to be moved or removed.

Removing the link is the right choice because users mostly manually
uncheck boxes rather than bulk unchecking them using the link. This is
based on the user research I have observed from finder-frontend. I'm
very interested to know if other groups using widget agree or have
contradictory findings (@rbouton, @quis, @tombye, @edds any reckons?)
Happy to revisit the design if the decision to remove is controversial.