Create report a problem multivariate test
If a page is being tested (currently set to all pages), then create a
standard test named `report-a-problem-redesign-ab-test`, with two
variants: the original design and the new design.

Refactor tests to ensure that the expected behaviours works in all
three variants: no test, control and new design. Move standard tests
into a helper that can be called in different contexts.