Add checked count to option-select label
One of the side effects of the closing link was that it served as an
indication that the list of checkboxes had a checked box when the
option select is closed.

This is a useful side effect from it's intended purpose.

The proposed fix here is to add an additional line under the label with
the selected count. This is an approach used by the digital marketplace
who are using a version of this widget.