Remove robots blocking of /apply-for-a-licence
This was added in 2012, and /apply-for-a-licence doesn't now exist
(or rather, it's a redirect to /licence-finder).

Also, robots.txt exclusions are prefixes, so this line was also banning
https://www.gov.uk/apply-for-a-licence-to-use-an-orphan-work from being
indexed.  In turn, this was causing search engines such as Google and
Bing to display a description for this page of "A description for this
result is not available because of this site's robots.txt".  Further,
Bing was attempting to guess a title from the URL and other off-page
information, and was picking "How to Apply for a License to Use an
Orphan Work"; note the incorrect spelling of "License".