Add close on load to govuk-option-select
In finder-frontend (which is the first client of this component) the
first option-select on the page load is open with has the rest closed.
This commit allows the client to pass in an argument to say "this
option should be closed on load". This does change the default
behaviour of the component, but as there are no clients for it yet
this shouldn't be a problem.

This commit:
- Adds a new optional parameter `closed_on_load`.
- Adds a fixture to the docs for `closed_on_load`
- Changes the view to append a data-attribute `data-closed-on-load`
which is then read by the JavaScript and used to determine if the
option-select should call `.close()` on itself at the end of
initialisation
- Refactors the specs for `.open()` to handle the new initialisation
which is open by default not closed by default
- Adds a test for the reading of `data-closed-on-load`