Header link should go to the homepage for the current env
In every GOV.UK environment, this link takes you to www.gov.uk. This seems
unhelpful when you're in an environment that isn't production as it takes you
to a different environment which can be hard to notice.

This already happens in preview, and now that we are adding GOV.UK Draft (which
is accessed on a different domain), there will be more users looking at the
frontend of GOV.UK in an environment that isn't "www.gov.uk".