Remove registration of /fonts route.
Nothing uses this route. The fonts are actually embedded in the
stylesheets using data-uri's. The fonts for IE8 are served through the
assets vhost under the /static/fonts path. There's nothing in this
application that serves anything from the /fonts path.

Another note, some of the handling for this route was removed in
https://github.gds/gds/puppet/pull/1153