Switch from `.live` to `.on`
`.live` is deprecated in jQuery 1.7

Weirdly, renaming live to on caused a parse error when running the
jasmine specs. The cause of this seems to be the above data attribute
with the value: 'javascript:void();’

Looking at the specs and survey I can’t see why this reset is needed,
and when removing it the tests continue to pass.