Prefix option select option id attr with key
If we have 2 option select components rendered on a page which have an
option in each with the same slug, these will be used as the id on the
input. This will then cause Nokogiri to throw an error in Slimmer and
not render the page.

This commit changes the option select component to prefix the option
value with the key (`key-slugified-value`).