Properly cleanup between survey JS tests
Tests begun to fail with jQuery 1.11.3

The existing `beforeEach` and `afterEach` weren’t adequately cleaning
up their HTML once a test was run. This led to multiple containers,
which led to the newer jQuery picking the first empty one, rather than
a later one containing the element we’re looking for.

* Declare a new $block at the beginning of each template
* Remove the $block at the end
* Simplify click handling on elements