Add analytics meta tag component
Create a component which takes a links hash object from a content-store
object and turns that into a collection of meta-tags which can then be
read by our analytics JavaScript. This information in then sent to
Google as a custom dimension.

The component doesn't actually render anything visual so doesn't have a
corresponding stylesheet.