Up the size of the checkbox on option selects
We’re upping the size of checkboxes generally across the site. Originally this code just used  a negative margin-left to pull the checkbox to the left, but this also pulls the following text regardless of padding. By floating it left the following text will respect the padding boundary set against the label.