Rich Govspeak with bold
As we need to have bold on certain pages of GOV.UK (the Highway Code)
this commit updates the Govspeak component to have extra CSS set by a
flag of `rich_govspeak`. This can be set by the rendering App.
Additional CSS could be added to this rich mode if we wish.

This has been discussed with both Product Managers, Content and
Designers and the approach I've followed was agreed upon.

[Ticket](https://trello.com/c/eYyOkNRR/131-highway-code-manual-has-styled-strong-tags).