Fix alignment of report-a-problem at certain widths
At a breakpoint below 1020 the report a problem link, and
form, weren't aligned to the grid of the rest of the page, and were
indented by an additional 30px on each side.

This was caused by `%site-width-container` being used on the report
a problem containers, within `#wrapper`, which already extends
`%site-width-container`, causing double margins.

As the report a problem elements already inside a page width container
it's same to remove the use of `%site-width-container` for them.