Fix analytics component fixtures
The component doc page was erroring as the default fixture doesn't
have enough data to actually render the component, and was erroring
on `content_item["links"] || {}`.

Replace the empty default with the first, that has the required
data.

Note; this component is unusual, in that it's highly coupled to the
structure of Content Items, and uses non-mustache compataible ERB
features, both of which kind of go against the principles of
components, and i'm not sure it's quite right. Or at least, if we're
going to do it, we should do it everywhere and save some pain.