Split Component docs into one file per component
This is an internal only change, and won't affect the response
the `govuk_component_docs` route returns - though it will enforce
an alphabetical order, which is a good default anyway

As we added more components, and fixtures, this file was growing
pretty big and hard to read/edit.

It's rare for someone to be looking at more than one component at
a time, the common case being working on fixtures for a single
component, so it makes sense to group the YML that way and keep
the files smaller.

If the number of fixtures continue to grow then it might be worth
splitting the fixtures out to one per file, grouped in a directory
by component.

Also updates the rails generator to use the file-per component
structure, which simplifies that code quite a bit too.