Remove unnecessary '---' from component YML docs
Despite Tommy's prodding in the previous PR (https://github.com/alphagov/static/pull/664#discussion_r42227207)
I messed this up, the `---` isn't required, and is only inserted
if you programatically write the YML files, which I did to convert
the single YML file to the many smaller ones.

It's not required, and the generator wasn't creating it anyway, so
we can drop it from the doc files, for simplicity's sake.