Fix Analytics Metadata component use with fixtures
Fixtures/the component guide generate an object with symbol keys,
as this is the convention of GOV.UK components, as we had confusion
between symbol and string keys, see:

  https://github.com/alphagov/static/pull/602

Analytics Metadata is an exception, because it's passed ContentItem
object directly from the `content_store` API, which will always
have string keys.

This also changes the structure of link sub-hashes to match the
structure expected by the component template, specifically, it
expects a field called `analytics_identifier` for a link, not the
link to have that as it's value.