Add some security to the web chat page messaging
The third parameter of postMessage should be a protocol and host that the message is scoped to. Previously this was set to the wildcard, but this commit changes it to use ENV vars to limit it. The exception is dev: on local static exposes www.dev.gov.uk as the domain but the actual domain is specific to the frontend calling it. For this scenario we drop back to *